Arxiv/requirements.txt
# Core UI
streamlit>=1.32.0
streamlit-extras>=0.4.0

# ArXiv data
arxiv>=2.1.0

# Embeddings & Vector Store
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0

# LLM Backends
ollama>=0.2.0
groq>=0.5.0
huggingface-hub>=0.22.0

# LangChain core + integrations
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.6
langchain-ollama>=0.1.0
langchain-huggingface>=0.0.3

# Data & ML
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
tiktoken>=0.6.0

# Visualizations
plotly>=5.20.0
wordcloud>=1.9.3
networkx>=3.3
matplotlib>=3.8.0
pyvis>=0.3.2
kaleido>=0.2.1

# NLP
transformers>=4.40.0
torch>=2.2.0

# Utils
python-dotenv>=1.0.1
tqdm>=4.66.2
requests>=2.31.0
aiohttp>=3.9.0
cachetools>=5.3.0
pyarrow>=15.0.0
